% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-filters.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply predefined color filter}
\usage{
apply_filter(
  nr,
  filter = c("1977", "aden", "brannan", "brooklyn", "clarendon", "earlybird", "gingham",
    "hudson", "inkwell", "kelvin", "lark", "lofi", "maven", "mayfair", "moon",
    "nashville", "reyes", "rise", "slumber", "stinson", "toaster", "valencia", "walden")
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{filter}{The name of the filter.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Applies one of the predefined color filters
that are ported from \href{https://github.com/ha-shine/rustagram}{Rustagram}.

The following filters are available:
\itemize{
\item 1977
\item aden
\item brannan
\item brooklyn
\item clarendon
\item earlybird
\item gingham
\item hudson
\item inkwell
\item kelvin
\item lark
\item lofi
\item maven
\item mayfair
\item moon
\item nashville
\item reyes
\item rise
\item slumber
\item stinson
\item toaster
\item valencia
\item walden
}
}
