% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrange_with}
\alias{arrange_with}
\title{Arrange a character column with patterns}
\usage{
arrange_with(tbl, col, patterns)
}
\arguments{
\item{tbl}{A data.frame.}

\item{col}{Column name.}

\item{patterns}{A character vector.}
}
\value{
data.frame.
}
\description{
Arrange a character column with patterns
}
