% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aznyan-package.R
\name{as_recordedplot}
\alias{as_recordedplot}
\title{Convert PNG image data into recorded plot}
\usage{
as_recordedplot(png)
}
\arguments{
\item{png}{A raw vector of PNG image.}
}
\value{
A recorded plot.
}
\description{
Convert PNG image data into recorded plot
}
