% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aznyan-package.R
\docType{package}
\name{aznyan-package}
\alias{aznyan}
\alias{aznyan-package}
\title{aznyan: Image Filters With 'OpenCV'}
\description{
Offers image filters wrapping 'OpenCV' \url{https://opencv.org/}, ported from \url{https://github.com/5PB-3-4/AviUtl_OpenCV_Scripts}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/aznyan}
  \item Report bugs at \url{https://github.com/paithiov909/aznyan/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item Aras Pranckevičius [copyright holder]
  \item Jeroen Ooms [copyright holder]
}

}
\keyword{internal}
