% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aznyan.R
\docType{package}
\name{aznyan}
\alias{aznyan}
\alias{aznyan-package}
\title{aznyan: An Annict GraphQL API Client}
\description{
Annict GraphQL API client for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/aznyan}
  \item Report bugs at \url{https://github.com/paithiov909/aznyan/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
\keyword{internal}
