% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aznyan-package.R
\name{aznyan_num_threads}
\alias{aznyan_num_threads}
\title{Get or set number of OpenCV threads}
\usage{
aznyan_num_threads(n)
}
\arguments{
\item{n}{An integer scalar larger than \code{1}.
If missing, returns the current number of threads.}
}
\value{
Current number of threads used by OpenCV.
}
\description{
Get or set number of OpenCV threads
}
\keyword{internal}
