% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{blend-mode}
\alias{blend-mode}
\alias{blend_alpha}
\alias{blend_darken}
\alias{blend_multiply}
\alias{blend_colorburn}
\alias{blend_lighten}
\alias{blend_screen}
\alias{blend_add}
\alias{blend_colordodge}
\alias{blend_hardlight}
\alias{blend_softlight}
\alias{blend_overlay}
\alias{blend_hardmix}
\alias{blend_linearlight}
\alias{blend_vividlight}
\alias{blend_pinlight}
\alias{blend_average}
\alias{blend_exclusion}
\alias{blend_difference}
\alias{blend_divide}
\alias{blend_subtract}
\alias{blend_luminosity}
\alias{blend_ghosting}
\title{Blend modes}
\usage{
blend_alpha(src, dst)

blend_darken(src, dst)

blend_multiply(src, dst)

blend_colorburn(src, dst)

blend_lighten(src, dst)

blend_screen(src, dst)

blend_add(src, dst)

blend_colordodge(src, dst)

blend_hardlight(src, dst)

blend_softlight(src, dst)

blend_overlay(src, dst)

blend_hardmix(src, dst)

blend_linearlight(src, dst)

blend_vividlight(src, dst)

blend_pinlight(src, dst)

blend_average(src, dst)

blend_exclusion(src, dst)

blend_difference(src, dst)

blend_divide(src, dst)

blend_subtract(src, dst)

blend_luminosity(src, dst)

blend_ghosting(src, dst)
}
\arguments{
\item{src, dst}{A \code{nativeRaster} object.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Blends two \code{nativeRaster} objects with the specified blend mode.
}
