% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aznyan-package.R
\docType{data}
\name{blue_noise_64x64}
\alias{blue_noise_64x64}
\title{Blue noise pattern}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 64 rows and 64 columns.
}
\usage{
blue_noise_64x64
}
\description{
Blue noise pattern of size 64x64
generated with \code{ambient::noise_blue(c(64, 64))}.

Range: \verb{[0, 1]}
}
\keyword{datasets}
