% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{blur}
\alias{blur}
\alias{median_blur}
\alias{box_blur}
\alias{gaussian_blur}
\title{Blur filters}
\usage{
median_blur(nr, ksize = 1)

box_blur(
  nr,
  box_w = 1,
  box_h = box_w,
  normalize = TRUE,
  border = c(3, 4, 0, 1, 2)
)

gaussian_blur(
  nr,
  box_w = 1,
  box_h = box_w,
  sigma_x = 0,
  sigma_y = sigma_x,
  border = c(3, 4, 0, 1, 2)
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{ksize}{An integer scalar specifying the kernel size.
The actual kernel size becomes \code{2 * ksize + 1}.}

\item{box_w}{An integer scalar controlling the half-size of the kernel in
the horizontal direction. The actual kernel width becomes \code{2 * box_w - 1}.}

\item{box_h}{An integer scalar controlling the half-size of the kernel in
the vertical direction. Defaults to \code{box_w}. The actual kernel height
becomes \code{2 * box_h - 1}.}

\item{normalize}{A logical scalar
specifying whether the kernel is normalized by its area or not.
Defaults to \code{TRUE}.}

\item{border}{An integer scalar specifying the border-handling mode.
One of \verb{0, 1, 2, 3, 4}, corresponding to OpenCV's border modes.}

\item{sigma_x, sigma_y}{A numeric scalar giving the standard deviation of the
Gaussian kernel along the x-axis. A value of \code{0} lets OpenCV compute it
automatically from the kernel size.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Blur filters
}
\details{
\code{border} corresponds to the OpenCV extrapolation types:
\enumerate{
\item cv::BORDER_CONSTANT
\item cv::BORDER_REPLICATE
\item cv::BORDER_REFLECT
\item cv::BORDER_REFLECT_101
\item cv::BORDER_ISOLATED
}
}
