% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{blurhash}
\alias{blurhash}
\title{BlurHash}
\usage{
blurhash(nr, x_comps = 6, y_comps = 6)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{x_comps, y_comps}{The number of x and y components.}
}
\value{
A \code{nativeRaster} object.
}
\description{
BlurHash
}
