% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{box_blur}
\alias{box_blur}
\title{Box blur}
\usage{
box_blur(
  png,
  box_w,
  box_h = box_w,
  normalize = TRUE,
  border = c(4, 0, 1, 2, 3)
)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{box_w}{The width of box.}

\item{box_h}{The height of box.}

\item{normalize}{Whether normalize.}

\item{border}{The type of pixel extrapolation method.}
}
\value{
A raw vector of PNG image.
}
\description{
Box blur
}
