% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{canny_filter}
\alias{canny_filter}
\title{Canny filter}
\usage{
canny_filter(
  png,
  asize = 3,
  balp = TRUE,
  use_rgb = TRUE,
  grad = FALSE,
  thres1 = 5,
  thres2 = 5
)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{asize}{Aparture size for Sobel operation.}

\item{balp}{Whether to remove black background.}

\item{use_rgb}{Whether to use RGB Canny filter.}

\item{grad}{Flag to be passed to \code{L2gradient}.}

\item{thres1}{The first threshold.}

\item{thres2}{The second threshold.}
}
\value{
A raw vector of PNG image.
}
\description{
Canny filter
}
