% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color-manip}
\alias{color-manip}
\alias{contrast}
\alias{brighten}
\alias{darken}
\alias{saturate}
\alias{grayscale}
\alias{sepia}
\title{Color manipulation}
\usage{
contrast(nr, intensity)

brighten(nr, intensity)

darken(nr, intensity)

saturate(nr, intensity)

grayscale(nr)

sepia(nr, intensity = 1, depth = 20)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{intensity}{A numeric scalar typically in range \verb{[0, 1]}.}

\item{depth}{An integer scalar.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Color manipulation
}
