% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-map.R
\name{color_map}
\alias{color_map}
\title{Apply a color map}
\usage{
color_map(nr, map = 0, use_hsv = FALSE, inverse = FALSE)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{map}{An integer scalar specifying the colormap ID.
Must be one of \verb{0–21}, corresponding to OpenCV's built-in color maps.}

\item{use_hsv}{A logical scalar.
If \code{TRUE}, the hue channel (HSV_FULL) is used as the input to the colormap;
if \code{FALSE}, a grayscale conversion is used instead.}

\item{inverse}{A logical scalar. If \code{TRUE}, the input channel is inverted
(\code{255 - value}) before applying the colormap.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Maps grayscale or hue values of a \code{nativeRaster} image to a predefined color
lookup table (OpenCV colormaps). The input can be derived either from the
grayscale intensity or from the hue channel in HSV space. The mapped values
replace the image's color channels while preserving the alpha channel.
}
\details{
\subsection{Options}{

\code{map} corresponds to:
\enumerate{
\item cv::COLORMAP_AUTUMN
\item cv::COLORMAP_BONE
\item cv::COLORMAP_JET
\item cv::COLORMAP_WINTER
\item cv::COLORMAP_RAINBOW
\item cv::COLORMAP_OCEAN
\item cv::COLORMAP_SUMMER
\item cv::COLORMAP_SPRING
\item cv::COLORMAP_COOL
\item cv::COLORMAP_HSV
\item cv::COLORMAP_PINK
\item cv::COLORMAP_HOT
\item cv::COLORMAP_PARULA
\item cv::COLORMAP_MAGMA
\item cv::COLORMAP_INFERNO
\item cv::COLORMAP_PLASMA
\item cv::COLORMAP_VIRIDIS
\item cv::COLORMAP_CIVIDIS
\item cv::COLORMAP_TWILIGHT
\item cv::COLORMAP_TWILIGHT_SHIFTED
\item cv::COLORMAP_TURBO
\item cv::COLORMAP_DEEPGREEN
}
}
}
