% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\docType{data}
\name{conversion-mat}
\alias{conversion-mat}
\alias{M_Rec709_to_sRGB}
\alias{M_sRGB_to_Rec709}
\title{Matrices to convert between color spaces}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.

An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\usage{
M_Rec709_to_sRGB

M_sRGB_to_Rec709
}
\description{
Matrices to convert between color spaces
}
\keyword{datasets}
