% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{detail_enchance}
\alias{detail_enchance}
\title{Detail enhancement filter}
\usage{
detail_enchance(nr, sgmS = 10, sgmR = 0.15)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{sgmS}{A numeric scalar giving the spatial sigma used for edge-preserving
filtering. Larger values increase the spatial smoothing radius.
Valid range: \verb{0–200}.}

\item{sgmR}{A numeric scalar giving the range sigma controlling how strongly
edges are preserved based on color differences.
Valid range: \verb{0–1}.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Applies OpenCV's \code{detailEnhance} to a \code{nativeRaster} image.
This filter enhances fine structures such as
edges and textures while preserving the overall appearance of the image.
}
