% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{diffusion_filter}
\alias{diffusion_filter}
\title{Diffusion filter}
\usage{
diffusion_filter(
  png,
  factor = 5,
  offset = 0.1,
  iter = 3,
  gamma = 1.3,
  sigma = 2
)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{factor}{The factor of diffusion.}

\item{offset}{The offset of diffusion.}

\item{iter}{The iteration of diffusion.}

\item{gamma}{The gamma of diffusion.}

\item{sigma}{The sigma of diffusion.}
}
\value{
A raw vector of PNG image.
}
\description{
Diffusion filter
}
