% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{edge-extract}
\alias{edge-extract}
\alias{sobel_filter}
\alias{laplacian_filter}
\alias{canny_filter}
\title{Edge extraction}
\usage{
sobel_filter(
  nr,
  ksize = 3,
  balp = TRUE,
  use_rgb = TRUE,
  border = c(3, 4, 0, 1, 2),
  dx = 1,
  dy = dx,
  scale = 1,
  delta = 0
)

laplacian_filter(
  nr,
  ksize = 3,
  balp = TRUE,
  use_rgb = TRUE,
  border = c(3, 4, 0, 1, 2),
  scale = 1,
  delta = 0
)

canny_filter(
  nr,
  asize = 2,
  balp = FALSE,
  use_rgb = TRUE,
  thres1 = 100,
  thres2 = 200,
  grad = TRUE
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{ksize}{An integer scalar.
The size of kernel.}

\item{balp}{A logical scalar.
If \code{TRUE}, resets transparency by thresholding.}

\item{use_rgb}{A logical scalar.
If \code{TRUE}, extracts edges separately for each color channel
and combines them.}

\item{border}{An integer scalar.
The type of pixel extrapolation method.}

\item{dx, dy}{An integer scalar.
Order of derivative for Sobel filter.}

\item{scale, delta}{A numeric scalar.
Optional parameters.}

\item{asize}{An integer scalar.
The size of aperture for Sobel operation.}

\item{thres1, thres2}{A numeric scalar.
Thresholds for the hysteresis procedure.}

\item{grad}{A logical scalar.
Flag to be passed to \code{L2gradient}.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Edge extraction
}
\details{
\code{border} refers to:
\enumerate{
\item cv::BORDER_CONSTANT
\item cv::BORDER_REPLICATE
\item cv::BORDER_REFLECT
\item cv::BORDER_REFLECT_101
\item cv::BORDER_ISOLATED
}
}
