% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-manip.R
\name{fill_with}
\alias{fill_with}
\title{Create a native raster filled with a color}
\usage{
fill_with(color, width, height)
}
\arguments{
\item{color}{Color name or hex code.}

\item{width, height}{A positive integer scalar.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Create a native raster filled with a color
}
