% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{gaussian_blur}
\alias{gaussian_blur}
\title{Gaussian blur}
\usage{
gaussian_blur(
  png,
  box_w,
  box_h = box_w,
  sigma_x,
  sigma_y = sigma_x,
  border = c(4, 0, 1, 2, 3)
)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{box_w}{The width of box.}

\item{box_h}{The height of box.}

\item{sigma_x}{The sigma of x.}

\item{sigma_y}{The sigma of y.}

\item{border}{The type of pixel extrapolation method.}
}
\value{
A raw vector of PNG image.
}
\description{
Gaussian blur
}
