% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lyrics.R
\name{get_lyrics}
\alias{get_lyrics}
\title{Scrape lyrics from list}
\usage{
get_lyrics(df, file, links = "link")
}
\arguments{
\item{df}{A tibble that comes of \code{get_lyrics_list}.}

\item{file}{String; file name to append lyrircs.}

\item{links}{String; column name of lyrics links.}
}
\value{
\code{file} is returned invisibly.
}
\description{
Scrape lyrics from list
}
\examples{
\dontrun{
csv_file <-
 aznyan::get_lyrics_list("23729") |>
 aznyan::get_lyrics("23729.csv")
tbl <-
 readr::read_csv(csv_file, col_names = F, col_types = "cccc___cDn") |>
 dplyr::rename(
   title = X1,
   artist = X2,
   lyricist = X3,
   composer = X4,
   text = X8,
   released = X9,
   page_view = X10
 )
}
}
