% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics_list.R
\name{get_lyrics_list}
\alias{get_lyrics_list}
\title{Scrape table of lyrics list}
\usage{
get_lyrics_list(id, type = c("artist", "lyricist", "composer"))
}
\arguments{
\item{id}{String; substring xxx of 'https://www.uta-net.com/:type:/xxx/'.}

\item{type}{String; one of "artist", "lyricist", or "composer".}
}
\value{
tibble.
}
\description{
Scrape table of lyrics list
}
