% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{laplacian_filter}
\alias{laplacian_filter}
\title{Laplacian filter}
\usage{
laplacian_filter(
  nr,
  ksize = 3,
  balp = TRUE,
  border = c(3, 4, 0, 1, 2),
  use_rgb = TRUE,
  scale = 1,
  delta = 0
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{ksize}{The size of kernel.}

\item{balp}{Whether to remove black background.}

\item{border}{The type of pixel extrapolation method.}

\item{use_rgb}{Whether to use RGB Laplacian filter.}

\item{scale}{The scale of Laplacian filter.}

\item{delta}{The delta of Laplacian filter.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Laplacian filter
}
