% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{mean_shift}
\alias{mean_shift}
\title{Mean shift filter}
\usage{
mean_shift(nr, sp = 10, sr = 30, max_level = 1)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{sp, sr}{The parameters of mean shift.}

\item{max_level}{The maximum level of mean shift.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Mean shift filter
}
