% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{median_blur}
\alias{median_blur}
\title{Median blur}
\usage{
median_blur(nr, ksize = 1)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{ksize}{The size of kernel.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Median blur
}
