% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pack-unpack}
\alias{pack-unpack}
\alias{pack_color}
\alias{unpack_color}
\title{Pack and unpack RGBA values}
\usage{
pack_color(r, g, b, a)

unpack_color(x)
}
\arguments{
\item{r, g, b, a}{Numeric vectors of equal length in range \verb{[0, 255]}.}

\item{x}{An integer vector that contains native packed integers.}
}
\value{
Integers.
}
\description{
Pack and unpack RGBA values
}
