% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_to_jdate}
\alias{parse_to_jdate}
\title{Parse dates to Japanese dates}
\usage{
parse_to_jdate(date, format = enc2utf8("Gy年M月d日"))
}
\arguments{
\item{date}{Dates.}

\item{format}{String.}
}
\value{
a chacter vector.
}
\description{
Parse dates to Japanese dates
}
