% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{read_cube}
\alias{read_cube}
\title{Read a 3D LUT file}
\usage{
read_cube(lut, verbose = TRUE)
}
\arguments{
\item{lut}{The path to cube LUT file.}

\item{verbose}{Whether to print debug messages.}
}
\value{
A data frame with columns \code{x}, \code{y} and \code{z}.
}
\description{
Read a 3D LUT file
}
