% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{rec709}
\alias{rec709}
\alias{decode_rec709}
\alias{encode_rec709}
\title{Gamma and inverse gamma in 'Rec.709' color space}
\usage{
decode_rec709(x)

encode_rec709(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
Gamma and inverse gamma in 'Rec.709' color space
}
