% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resample}
\alias{resample}
\title{Resample image}
\usage{
resample(
  nr,
  wh = c(0.2, 0.2),
  resize_mode1 = c(1, 2, 3, 4, 5, 6, 0),
  resize_mode2 = c(1, 2, 3, 4, 5, 6, 0)
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{wh}{The width and height coef for reduction.}

\item{resize_mode1}{The resize mode for reduction.}

\item{resize_mode2}{The resize mode for expansion.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Resample image
}
