% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-manip.R
\name{rgb-hls}
\alias{rgb-hls}
\alias{rgb2hls}
\alias{hls2rgb}
\title{Conversion between RGB and HLS color spaces}
\usage{
rgb2hls(x)

hls2rgb(x)
}
\arguments{
\item{x}{An integer matrix with 3 rows.}
}
\value{
An integer matrix of the same size as \code{x}
}
\description{
Conversion between RGB and HLS color spaces
}
