% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{rgb-hsl}
\alias{rgb-hsl}
\alias{rgb2hls}
\alias{hls2rgb}
\title{Conversion between RGB and HSL}
\usage{
rgb2hls(x)

hls2rgb(x)
}
\arguments{
\item{x}{An integer matrix with 3 rows.}
}
\value{
An integer matrix of the same size as \code{x}
}
\description{
Conversion between RGB and HSL
}
