% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics_list.R
\name{search_lyrics_list}
\alias{search_lyrics_list}
\title{Search lyrics list by keyword}
\usage{
search_lyrics_list(keyword, sort = c("new", "popular", "title", "artist"))
}
\arguments{
\item{keyword}{String; search phrase.}

\item{sort}{String; one of "new", "popular", "title", or "artist".}
}
\value{
tibble
}
\description{
Search lyrics list by keyword
}
