% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort.nativeRaster}
\alias{sort.nativeRaster}
\title{Sort method for native raster images}
\usage{
\method{sort}{nativeRaster}(
  x,
  decreasing = FALSE,
  na.last = FALSE,
  by = c("luma", "blue", "green", "red", "hue", "luminance", "saturation", "value"),
  ...
)
}
\arguments{
\item{x}{for \code{sort} an \R object with a class or a numeric,
    complex, character or logical vector.  For \code{sort.int}, a
    numeric, complex, character or logical vector, or a factor.}

\item{decreasing}{logical.  Should the sort be increasing or decreasing?
    Not available for partial sorting.}

\item{na.last}{This argument is ignored because \code{NA_integer_} is a valid pixel value
in native raster images.}

\item{by}{A string specifying the property to sort by.}

\item{...}{arguments to be passed to or from methods or (for the
    default methods and objects without a class) to \code{sort.int}.}
}
\description{
Sort method for native raster images
}
\keyword{internal}
