% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{stylize}
\alias{stylize}
\title{Stylization filter}
\usage{
stylize(nr, sgmS = 60, sgmR = 0.44)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{sgmS}{A numeric scalar giving the spatial sigma controlling the
smoothness of the stylization process.
Valid range: \verb{0–200}.}

\item{sgmR}{A numeric scalar giving the range sigma determining sensitivity
to color differences during abstraction.
Valid range: \verb{0–1}.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Applies a stylization effect to a \code{nativeRaster} image.
This filter combines edge-aware smoothing
and abstraction to produce a painterly,
cartoon-like appearance while preserving prominent structures.
}
