% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{swap_channels}
\alias{swap_channels}
\title{Swap channels}
\usage{
swap_channels(png, from = c(0, 1, 2, 3), to = c(1, 2, 0, 3))
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{from}{The channels to swap from.}

\item{to}{The channels to swap to.}
}
\value{
A raw vector of PNG image.
}
\description{
Swap channels
}
