% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{times}
\alias{times}
\title{Call a function n times}
\usage{
times(fn, n = 2L, ..., .walk = FALSE)
}
\arguments{
\item{fn}{A function.}

\item{n}{Integer.}

\item{...}{Other arguments are passed to function fn.}

\item{.walk}{If supplied \code{TRUE}, executes the function
inside \code{purrr::walk} rather than \code{purrr::map}.}
}
\description{
Simply executes a function specified times inside \code{purrr::map}.
}
