% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warp_perspective}
\alias{warp_perspective}
\title{Warp a \code{nativeRaster} image using perspective transformation}
\usage{
warp_perspective(nr, mat = diag(1, 3), border = c(3, 4, 0, 1, 2))
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{mat}{A 3x3 matrix specifying the perspective transformation.}

\item{border}{An integer scalar selecting pixel extrapolation method.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Warps an image using a perspective transformation,
keeping the same dimensions as the input image.
}
\details{
\code{border} corresponds to the OpenCV extrapolation types:
\enumerate{
\item cv::BORDER_CONSTANT
\item cv::BORDER_REPLICATE
\item cv::BORDER_REFLECT
\item cv::BORDER_WRAP
\item cv::BORDER_REFLECT_101
}
}
