% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut-io.R
\name{write_cube}
\alias{write_cube}
\alias{write_smcube}
\title{Write a 3D LUT file}
\usage{
write_cube(
  x,
  filename = tempfile(fileext = ".cube"),
  title = "",
  domain_min = c(0, 0, 0),
  domain_max = c(1, 1, 1)
)

write_smcube(
  x,
  filename,
  title = "",
  domain_min = c(0, 0, 0),
  domain_max = c(1, 1, 1)
)
}
\arguments{
\item{x}{A data frame with 3 numeric columns.}

\item{filename}{The path to save the LUT file.
It must end with \code{.cube} or \code{.smcube} respectively.}

\item{title}{\code{TITLE} for the LUT.}

\item{domain_min}{\code{DOMAIN_MIN} for the LUT.}

\item{domain_max}{\code{DOMAIN_MAX} for the LUT.}
}
\value{
The path to the saved LUT file is invisibly returned.
}
\description{
Saves 3 numeric columns of a data frame to a LUT file.
This function supports 3D, 3-channel LUTs only.
}
