% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xtabs_tidy}
\alias{xtabs_tidy}
\title{Xtabs with tidy selecting}
\usage{
xtabs_tidy(tbl, ..., .by = NULL)
}
\arguments{
\item{tbl}{A data.frame}

\item{...}{Column names go to right side of formula (tidy-selection).}

\item{.by}{Column name goes to left side of formula.}
}
\value{
a contingency table in array representation of S3 class
\code{c("xtabs", "table")}.
}
\description{
Xtabs with tidy selecting
}
