% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aznyan-package.R
\name{as_recordedplot}
\alias{as_recordedplot}
\title{Convert image data into a recorded plot}
\usage{
as_recordedplot(nr)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}
}
\value{
A recorded plot is invisibly returned.
}
\description{
Convert image data into a recorded plot
}
\keyword{internal}
