% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{bilateral_filter}
\alias{bilateral_filter}
\title{Bilateral filter}
\usage{
bilateral_filter(
  nr,
  d = 5,
  sigmacolor = 1,
  sigmaspace = 1,
  border = c(3, 4, 0, 1, 2),
  alphasync = TRUE
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{d}{The size of kernel.}

\item{sigmacolor}{The sigma of color.}

\item{sigmaspace}{The sigma of space.}

\item{border}{The type of pixel extrapolation method.}

\item{alphasync}{Whether sync alpha.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Bilateral filter
}
