% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{blur}
\alias{blur}
\alias{median_blur}
\alias{box_blur}
\alias{gaussian_blur}
\title{Blur filters}
\usage{
median_blur(nr, ksize = 1)

box_blur(
  nr,
  box_w = 1,
  box_h = box_w,
  normalize = TRUE,
  border = c(3, 4, 0, 1, 2)
)

gaussian_blur(
  nr,
  box_w = 1,
  box_h = box_w,
  sigma_x = 0,
  sigma_y = sigma_x,
  border = c(3, 4, 0, 1, 2)
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{ksize, box_w, box_h}{A numeric scalar. The kernel size.}

\item{normalize}{A logical scalar.
Whether the kernel is normalized by its area or not.}

\item{border}{An integer scalar.
The type of pixel extrapolation method.}

\item{sigma_x, sigma_y}{A numeric scalar.
The standard deviation of the Gaussian kernel in X and Y direction.
If both are 0, they are computed from \code{ksize}.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Blur filters
}
\details{
\code{border} refers to:
\enumerate{
\item cv::BORDER_CONSTANT
\item cv::BORDER_REPLICATE
\item cv::BORDER_REFLECT
\item cv::BORDER_REFLECT_101
\item cv::BORDER_ISOLATED
}
}
