% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{box_blur}
\alias{box_blur}
\title{Box blur}
\usage{
box_blur(
  nr,
  box_w = 1,
  box_h = box_w,
  normalize = TRUE,
  border = c(3, 4, 0, 1, 2)
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{box_w}{The width of box.}

\item{box_h}{The height of box.}

\item{normalize}{Whether normalize.}

\item{border}{The type of pixel extrapolation method.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Box blur
}
