% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{canny_filter}
\alias{canny_filter}
\title{Canny filter}
\usage{
canny_filter(
  nr,
  asize = 2,
  balp = FALSE,
  use_rgb = TRUE,
  grad = TRUE,
  thres1 = 100,
  thres2 = 200
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{asize}{Aparture size for Sobel operation.}

\item{balp}{Whether to remove black background.}

\item{use_rgb}{Whether to use RGB Canny filter.}

\item{grad}{Flag to be passed to \code{L2gradient}.}

\item{thres1}{The first threshold.}

\item{thres2}{The second threshold.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Canny filter
}
