% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-manip.R
\name{color-manip}
\alias{color-manip}
\alias{unpremul}
\alias{set_matte}
\alias{reset_alpha}
\alias{hue_rotate}
\alias{contrast}
\alias{brighten}
\alias{saturate}
\alias{grayscale}
\alias{sepia}
\title{Color manipulation}
\usage{
unpremul(nr, max = 255L)

set_matte(nr, color = "green")

reset_alpha(nr, alpha = 1)

hue_rotate(nr, rad)

contrast(nr, intensity)

brighten(nr, intensity)

saturate(nr, intensity)

grayscale(nr)

sepia(nr, intensity = 1, depth = 20)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{max}{An integer scalar. The maximum value of the color code.}

\item{color}{A character string; the color to be set.}

\item{alpha}{A numeric scalar in range \verb{[0, 1]}.
The alpha value to be reset for transparency.}

\item{rad}{A numeric scalar. The rotation angle in radians.}

\item{intensity}{A numeric scalar.}

\item{depth}{An integer scalar.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Color manipulation
}
