% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-map.R
\name{color_map}
\alias{color_map}
\title{Apply colormap}
\usage{
color_map(nr, map = 0, use_hsv = FALSE, inverse = FALSE)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{map}{An integer scalar.}

\item{use_hsv}{A logical scalar.
If \code{TRUE}, the map source is taken from the hue channel
converted with \code{cv::COLOR_BGR2HSV_FULL}.}

\item{inverse}{A logical scalar.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Applies a GNU Octave/MATLAB equivalent colormap to an image.

\code{map} refers to:
\enumerate{
\item cv::COLORMAP_AUTUMN
\item cv::COLORMAP_BONE
\item cv::COLORMAP_JET
\item cv::COLORMAP_WINTER
\item cv::COLORMAP_RAINBOW
\item cv::COLORMAP_OCEAN
\item cv::COLORMAP_SUMMER
\item cv::COLORMAP_SPRING
\item cv::COLORMAP_COOL
\item cv::COLORMAP_HSV
\item cv::COLORMAP_PINK
\item cv::COLORMAP_HOT
\item cv::COLORMAP_PARULA
\item cv::COLORMAP_MAGMA
\item cv::COLORMAP_INFERNO
\item cv::COLORMAP_PLASMA
\item cv::COLORMAP_VIRIDIS
\item cv::COLORMAP_CIVIDIS
\item cv::COLORMAP_TWILIGHT
\item cv::COLORMAP_TWILIGHT_SHIFTED
\item cv::COLORMAP_TURBO
\item cv::COLORMAP_DEEPGREEN
}
}
