% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{convolve}
\alias{convolve}
\title{Convolution with a custom kernel}
\usage{
convolve(
  nr,
  kernel = matrix(c(0, -1, 0, -1, 5, -1, 0, -1, 0), 3, 3),
  border = c(3, 4, 0, 1, 2),
  alphasync = FALSE
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{kernel}{A numeric matrix that represents a kernel.}

\item{border}{An integer scalar.
The type of pixel extrapolation method.}

\item{alphasync}{A logical scalar.
If \code{TRUE}, filtering is applied separately to the alpha channel.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Convolution with a custom kernel
}
\details{
\code{border} refers to:
\enumerate{
\item cv::BORDER_CONSTANT
\item cv::BORDER_REPLICATE
\item cv::BORDER_REFLECT
\item cv::BORDER_REFLECT_101
\item cv::BORDER_ISOLATED
}
}
