% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{diffusion_filter}
\alias{diffusion_filter}
\title{Diffusion filter}
\usage{
diffusion_filter(
  nr,
  factor = 5,
  offset = 0.1,
  iter = 3,
  gamma = 1.3,
  sigma = 2
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{factor, offset}{The parameters of diffusion.}

\item{iter}{The iteration of diffusion step.}

\item{gamma}{A numeric scalar. The gamma value for preprocessing.}

\item{sigma}{A numeric scalar.
The sigma of gaussian blur for preprocessing.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Diffusion filter
}
