% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aznyan.R
\name{get_lyrics}
\alias{get_lyrics}
\title{Scrape lyrics from list}
\usage{
get_lyrics(df, links = "link")
}
\arguments{
\item{df}{A tibble that comes of \code{get_lyrics_list}.}

\item{links}{String; column name of lyrics links.}
}
\value{
tibble.
}
\description{
Scrape lyrics from list
}
