% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lyrics.R
\name{get_lyrics}
\alias{get_lyrics}
\title{Scrape lyrics from list}
\usage{
get_lyrics(df, file, links = "link")
}
\arguments{
\item{df}{A tibble that comes of \code{get_lyrics_list}.}

\item{file}{String; file name to append lyrircs.}

\item{links}{String; column name of lyrics links.}
}
\value{
\code{file} is returned invisibly.
}
\description{
Scrape lyrics from list
}
