% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blur.R
\name{median_blur}
\alias{median_blur}
\title{Median blur}
\usage{
median_blur(png, ksize = 1)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{ksize}{The size of kernel.}
}
\value{
A raw vector of PNG image.
}
\description{
Median blur
}
