% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{morphology}
\alias{morphology}
\title{Morphology}
\usage{
morphology(
  png,
  ksize = c(2, 2, 2),
  ktype = c(0, 1, 2),
  mode = c(0, 1, 2, 3, 4, 5, 6, 7),
  border = c(3, 4, 0, 1, 2),
  iterations = 1,
  alphasync = TRUE,
  use_rgb = TRUE,
  anchor = c(-1, -1)
)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{ksize}{The size of kernel.}

\item{ktype}{The type of kernel.}

\item{mode}{The mode of morphology.}

\item{border}{The type of pixel extrapolation method.}

\item{iterations}{The iteration of morphology.}

\item{alphasync}{Whether sync alpha.}

\item{use_rgb}{Whether to use RGB morphology.}

\item{anchor}{The anchor of morphology.}
}
\value{
A raw vector of PNG image.
}
\description{
Morphology
}
\details{
\code{mode} is an integer scalar in range \verb{[0, 7]} corresponding to:
\enumerate{
\item cv::MORPH_ERODE
\item cv::MORPH_DILATE
\item cv::MORPH_OPEN
\item cv::MORPH_CLOSE
\item cv::MORPH_GRADIENT
\item cv::MORPH_TOPHAT
\item cv::MORPH_BLACKHAT
\item cv::MORPH_HITMISS
}
}
