% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-filters.R
\name{premul}
\alias{premul}
\title{Create color code while premultiplying alpha}
\usage{
premul(r, g, b, a, max = 255)
}
\arguments{
\item{r, g, b, a}{Numeric scalars.}

\item{max}{A numeric scalar.}
}
\value{
A color code.
}
\description{
Create color code while premultiplying alpha
}
