% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{preserve_edge}
\alias{preserve_edge}
\title{Preserve-edge filter}
\usage{
preserve_edge(nr, sgmS = 600, sgmR = 40, recursive = TRUE)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{sgmS, sgmR}{The parameters of edge preserving filter.}

\item{recursive}{Whether to use recursive mode.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Preserve-edge filter
}
