% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{preserve_edge}
\alias{preserve_edge}
\title{Preserve-edge filter}
\usage{
preserve_edge(png, sgmS = 600, sgmR = 40, recursive = TRUE)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{sgmS, sgmR}{The parameters of edge preserving filter.}

\item{recursive}{Whether to use recursive mode.}
}
\value{
A raw vector of PNG image.
}
\description{
Preserve-edge filter
}
