% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resample}
\alias{resample}
\title{Resample image}
\usage{
resample(
  png,
  wh = c(0.2, 0.2),
  resize_mode1 = c(1, 2, 3, 4, 5, 6, 0),
  resize_mode2 = c(1, 2, 3, 4, 5, 6, 0)
)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{wh}{The width and height coef for reduction.}

\item{resize_mode1}{The resize mode for reduction.}

\item{resize_mode2}{The resize mode for expansion.}
}
\value{
A raw vector of PNG image.
}
\description{
Resample image
}
