% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{resize}
\alias{resize}
\alias{resample}
\title{Resize and resample}
\usage{
resize(
  nr,
  wh = c(1, 1),
  resize_mode = c(1, 2, 3, 4, 5, 6, 0),
  set_size = FALSE
)

resample(
  nr,
  wh = c(0.2, 0.2),
  resize_mode1 = c(1, 2, 3, 4, 5, 6, 0),
  resize_mode2 = c(1, 2, 3, 4, 5, 6, 0)
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{wh}{A numeric vector of length 2. The width and height coef for reduction.}

\item{resize_mode, resize_mode1, resize_mode2}{An integer scalar.
The resize mode. For \code{resample} \code{1} is for reduction and \code{2} is for expansion.}

\item{set_size}{A logical scalar.
If \code{TRUE}, \code{wh} is treated as actual width and height instead of coef.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Resize and resample
}
