% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{sobel_filter}
\alias{sobel_filter}
\title{Sobel filter}
\usage{
sobel_filter(
  nr,
  ksize = 3,
  balp = TRUE,
  dx = 1,
  dy = dx,
  border = c(3, 4, 0, 1, 2),
  use_rgb = TRUE,
  scale = 1,
  delta = 0
)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{ksize}{The size of kernel.}

\item{balp}{Whether to remove black background.}

\item{dx}{The direction of Sobel filter.}

\item{dy}{The direction of Sobel filter.}

\item{border}{The type of pixel extrapolation method.}

\item{use_rgb}{Whether to use RGB Sobel filter.}

\item{scale}{The scale of Sobel filter.}

\item{delta}{The delta of Sobel filter.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Sobel filter
}
