% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{swap_channels}
\alias{swap_channels}
\title{Swap channels}
\usage{
swap_channels(nr, from = c(0, 1, 2, 3), to = c(1, 2, 0, 3))
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{from, to}{An integer vector of length 4. The channels to swap from and to.}
}
\value{
A \code{nativeRaster} object.
}
\description{
Swap channels
}
