% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thres.R
\name{thres}
\alias{thres}
\alias{adpthres}
\title{Image thresholding}
\usage{
thres(png, threshold = 100, maxv = 255, mode = c(0, 1, 2, 3, 4, 5, 6))

adpthres(png, maxv = 255, bsize = 1, C = 5, mode = c(0, 1), invert = FALSE)
}
\arguments{
\item{png}{A raw vector of PNG image.}

\item{threshold}{The threshold value.}

\item{maxv}{The maximum value.}

\item{mode}{The mode of thresholding.}

\item{bsize}{The block size.}

\item{C}{The C value for adaptive thresholding.}

\item{invert}{Whether to invert the result in adaptive thresholding.}
}
\value{
A raw vector of PNG image.
}
\description{
Image thresholding
}
